% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_latex.R
\name{theme_latex}
\alias{theme_latex}
\title{Add LaTeX theme to ggplot object}
\usage{
theme_latex(font = "lmroman", base_size = 12)
}
\arguments{
\item{font}{character; "lmroman" by default}

\item{base_size}{numeric; 12 by default}
}
\description{
This function changes ggplot object theme to LaTeX and color blind friendly style.
}
\examples{
library(ggplot2)
library(ltxplot)
load_theme_ltx()
df <- mtcars
df$carb <- as.factor(df$carb)
ggplot(df, aes(x=carb, fill = carb))+
  geom_bar(stat= "count", width=0.7)+
  labs(title = 'Number of cars with different number of carburetors',
       x = "Number of carburetors",
       y = "Number of car models",
       tag = 'Figure 1',
       caption = "Data source: mtcars") +
  theme_latex(base_size = 14) +
  theme(legend.position = "none")
}
\keyword{theme_latex}
